requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
aiosmtplib==2.0.2
annotated-types==0.6.0
anyio==3.7.1
APScheduler==3.10.3
asgiref==3.6.0
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.2
black==23.3.0
blinker==1.7.0
boto3==1.28.40
botocore==1.31.40
bs4==0.0.1
build==1.1.1
cachetools==5.3.3
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==3.1.0
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.3
coloredlogs==15.0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.7
dataclasses-json==0.6.4
debugpy==1.6.7
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.1.0
diff-match-patch==20230430
dirtyjson==1.0.8
distlib==0.3.8
distro==1.9.0
Django==4.1.7
django-apscheduler==0.6.2
django-cors-headers==3.14.0
django-crontab==0.7.1
django-filter==23.1
django-import-export==3.3.1
django-rest-swagger==2.2.0
django-ses==3.5.0
django-smtp-ssl==1.0
django-suit==0.2.28
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
dnspython==2.6.1
drf-yasg==1.21.7
email_validator==2.1.1
et-xmlfile==1.1.0
faiss-cpu==1.7.4
fastapi==0.110.0
fastapi-mail==1.4.1
filelock==3.13.1
Flask==3.0.2
Flask-Cors==4.0.0
flatbuffers==23.5.26
frozenlist==1.4.1
fsspec==2024.2.0
google-auth==2.28.1
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.62.0
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.3
httptools==0.6.1
httpx==0.26.0
huggingface-hub==0.21.3
humanfriendly==10.0
idna==3.4
import-export==0.3.1
importlib-metadata==6.11.0
importlib_resources==6.1.2
inflection==0.5.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
kubernetes==29.0.0
langchain==0.1.11
langchain-community==0.0.28
langchain-core==0.1.31
langchain-openai==0.0.8
langchain-text-splitters==0.0.1
langsmith==0.1.22
llama-index-agent-openai==0.1.6
llama-index-core==0.10.20.post2
llama-index-embeddings-openai==0.1.6
llama-index-llms-openai==0.1.12
llama-index-program-openai==0.1.4
llama-index-readers-wikipedia==0.1.3
llama-index-vector-stores-chroma==0.1.6
llama-index-vector-stores-pinecone==0.1.4
llamaindex-py-client==0.1.13
Mako==1.3.2
Markdown==3.4.3
MarkupPy==1.14
MarkupSafe==2.1.3
marshmallow==3.20.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numpy==1.25.2
oauthlib==3.2.2
odfpy==1.4.1
onnxruntime==1.17.1
openai==1.14.1
openapi-codec==1.3.2
openpyxl==3.1.2
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-asgi==0.44b0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
opentelemetry-util-http==0.44b0
orjson==3.9.15
overrides==7.7.0
packaging==23.2
pandas==2.2.1
parsimonious==0.10.0
passlib==1.7.4
pathspec==0.11.1
pillow==10.2.0
pinecone-client==3.1.0
pipenv==2023.12.1
platformdirs==4.2.0
posthog==3.5.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pulsar-client==3.4.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
PyJWT==1.7.1
PyMySQL==1.1.0
pypdf==4.0.2
PyPika==0.48.9
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-http-client==3.3.7
python-multipart==0.0.9
python-slugify==8.0.4
pytz==2023.2
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.2
schedule==1.2.1
sendgrid==6.10.0
simplejson==3.19.2
singlestoredb==1.0.2
six==1.16.0
slugify==0.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.28
sqlparams==6.0.1
sqlparse==0.4.3
starkbank-ecdsa==2.2.0
starlette==0.36.3
stripe==7.3.0
sympy==1.12
sysrsync==1.1.0
tablib==3.5.0
tenacity==8.2.3
text-unidecode==1.3
tiktoken==0.6.0
tokenizers==0.15.2
toml==0.10.2
tqdm==4.66.2
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==1.26.15
uvicorn==0.27.1
uvloop==0.19.0
virtualenv==20.25.1
watchfiles==0.21.0
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.1
whitenoise==6.4.0
wrapt==1.16.0
xata==1.3.1
xlrd==2.0.1
xlwt==1.3.0
yarl==1.9.4
zipp==3.17.0
